change average_score datatype

+++ touristDB.php
@@ -63,7 +63,7 @@ $sql = "CREATE TABLE IF NOT EXISTS trips (
     image_url VARCHAR(255), /*行程封面 */
     userId CHAR(6), /*創建行程的使用者*/
     total_date INT, /*行程的總天數*/
-    average_score FLOAT,
+    average_score DECIMAL(2,1),
     rating_num INT
 )";
